# Optimistically expect simple strings for json