@rs 2017-06-23
That's a great optim. Could you please add a few comment explaining what you're doing (i.e. when you switch between the two technics) to make it easier to understand?